requirements.txt
crewai>=0.37.1
crewai-tools>=0.11.0
langchain>=0.2.7
langchain-community>=0.2.7
langchain-openai>=0.1.8
openai>=1.42.0
faiss-cpu>=1.8.0
duckduckgo-search>=6.1.3
streamlit>=1.36.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.1
requests>=2.32.0
pydantic>=2.7.0
